chore(clerk-react): Restrcture errors to separate module /errors module
