+++ packages/react/src/contexts/assertHelpers.ts
@@ -1,4 +1,4 @@
-import { noClerkProviderError, noGuaranteedLoadedError } from '../errors';
+import { noClerkProviderError, noGuaranteedLoadedError } from '../errors/messages';
 import { errorThrower } from '../utils';
 
 export function assertWrappedByClerkProvider(contextVal: unknown): asserts contextVal {
