+++ packages/react/src/index.ts
@@ -2,7 +2,13 @@ import './polyfills';
 
 export * from './components';
 export * from './contexts';
-export { EmailLinkErrorCode, isClerkAPIResponseError, isEmailLinkError, isKnownError, isMetamaskError } from './errors';
+export {
+  EmailLinkErrorCode,
+  isClerkAPIResponseError,
+  isEmailLinkError,
+  isKnownError,
+  isMetamaskError,
+} from './errors/messages';
 export * from './hooks';
 export { useEmailLink } from './hooks/useEmailLink';
 export type {
